refactoring parsing into separate file
